basename: handle empty arguments
check that the suffix is non-empty:
If strlen(argv[2]) < 1, the for loop would not run, and the status would
be uninitialized. This commit causes the suffix to be ignored if it is
empty.

print empty line when input name is empty:
If basename(3) receives an empty string or a null pointer, it returns
the string ".", but we want basename(1) to just output an empty line
when the input is empty.