Use a new build system for everything, and revamp directories.
Previously there was quite a bit of sed(1)'ing going on on each source
file, and a bit of copying them to box_tmp. Now we use preprocessor
directives instead, things should be much cleaner. And since it's
a pain to prefix strings, move all programs out of subdirs.

I've also clarified the Makefile system a bit. I'm unsure how to make
standalone programs with POSIX Make, so currently we only generate
fases-in-a-box.